media: fix play error when play again after alloc codec fail once
PD# 158355

Problem:
This an acodec bug, When alloc component fail, it will signel error,
and then reclaim the last component alloc a new component again.
But this time mFatalError alse is true. So when dequeueBufferFromNativeWindow
it will "not dequeuing from native window due to fatal error",
and return NULL. It duo to video can not play.

Solution:
This fixed is when alloc a new component again success, set mFatalError to false

Verify:
verified by p212

Change-Id: I92c07da696cf0c4b328f01a4776901a02fac95bd